requirements.txt
aiogram>=3.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
apscheduler>=3.10.0
pydantic-settings>=2.0.0
aiohttp>=3.9.0
redis>=5.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
python-dateutil>=2.8.0
